pyproject.toml
[project]
name = "my-diary"
version = "0.1.0"
description = "Automatyczny skryba dzienny — zbiera aktywność z wielu źródeł i syntetyzuje dzienną notatkę"
readme = "README.md"
authors = [
    { name = "Maciej Kobuszewski", email = "maciej.kobuszewski@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "structlog>=24.0",
    "slack-sdk>=3.27",
    "aiohttp>=3.9",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.2",
]

[project.scripts]
my-diary = "my_diary.cli:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
